Move acyclic implementation into other class to shorten the cyclic implementation
