+++ reasoner/v4/nodes/ActorNode.java
@@ -1,10 +1,10 @@
-package com.vaticle.typedb.core.reasoner.v4;
+package com.vaticle.typedb.core.reasoner.v4.nodes;
 
 import com.vaticle.typedb.core.common.exception.TypeDBException;
 import com.vaticle.typedb.core.common.iterator.FunctionalIterator;
 import com.vaticle.typedb.core.common.iterator.Iterators;
 import com.vaticle.typedb.core.concurrent.actor.Actor;
-import com.vaticle.typedb.core.reasoner.v4.nodes.AnswerTable;
+import com.vaticle.typedb.core.reasoner.v4.Message;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -19,30 +19,12 @@ import java.util.function.Supplier;
 import static com.vaticle.typedb.core.common.exception.ErrorMessage.Internal.ILLEGAL_STATE;
 import static com.vaticle.typedb.core.common.exception.ErrorMessage.Internal.UNIMPLEMENTED;
 
-public abstract class ActorNode<NODE extends ActorNode<NODE>> extends Actor<NODE> {
+public abstract class ActorNode<NODE extends ActorNode<NODE>> extends AbstractAcyclicNode<NODE> {
 
-    private static final Logger LOG = LoggerFactory.getLogger(ActorNode.class);
-
-    protected final NodeRegistry nodeRegistry;
-    protected List<ActorNode.Port> ports;
-    private final Set<Port> activePorts;
-    private final Set<Port> pendingPorts;
-    protected final Integer nodeId;
-    protected final AnswerTable answerTable;
-
+    static final Logger LOG = LoggerFactory.getLogger(ActorNode.class);
 
     protected ActorNode(NodeRegistry nodeRegistry, Driver<NODE> driver, Supplier<String> debugName) {
-        super(driver, debugName);
-        this.nodeRegistry = nodeRegistry;
-        nodeId = nodeRegistry.nextNodeAge();
-        ports = new ArrayList<>();
-        activePorts = new HashSet<>();
-        pendingPorts = new HashSet<>();
-        answerTable = new AnswerTable();
-    }
-
-    protected void initialise() {
-
+        super(nodeRegistry, driver, debugName);
     }
 
     // TODO: Since port has the index in it, maybe we don't need index here?
@@ -63,30 +45,6 @@ public abstract class ActorNode<NODE extends ActorNode<NODE>> extends Actor<NODE
     public final void readAnswerAt(ActorNode.Port reader, int index) {
         throw TypeDBException.of(ILLEGAL_STATE);
     }
-    protected abstract void propagatePull(Port reader, int index);
-
-    public void receive(Port onPort, Message received) {
-        switch (received.type()) {
-            case ANSWER: {
-                handleAnswer(onPort, received.asAnswer());
-                break;
-            }
-            case CONCLUSION: {
-                handleConclusion(onPort, received.asConclusion());
-                break;
-            }
-            case SNAPSHOT: {
-                handleSnapshot(onPort);
-                break;
-            }
-            case DONE: {
-                handleDone(onPort);
-                break;
-            }
-            default:
-                throw TypeDBException.of(ILLEGAL_STATE);
-        }
-    }
 
     protected abstract void handleAnswer(Port onPort, Message.Answer answer);
 
@@ -104,48 +62,7 @@ public abstract class ActorNode<NODE extends ActorNode<NODE>> extends Actor<NODE
         Port port = new Port(this, remote);
         ports.add(port);
         activePorts.add(port);
-
         return port;
-    }
-
-    protected boolean allPortsDone() {
-        return activePorts.isEmpty() && pendingPorts.isEmpty();
-    }
-
-    protected boolean anyPortsActive() {
-        return !activePorts.isEmpty();
-    }
-
-    protected FunctionalIterator<ActorNode.Port> allPorts() {
-        return Iterators.iterate(ports);
-    }
-
-    // TODO: See if i can safely get recipient from port
-    public void send(ActorNode<?> recipient, ActorNode.Port recipientPort, Message message) {
-        assert recipientPort.remote == this;
-        recipient.driver().execute(actor -> actor.receiveOnPort(recipientPort, message));
-    }
-
-    protected void receiveOnPort(Port port, Message message) {
-        LOG.debug(port.owner() + " received " + message + " from " + port.remote());
-        port.recordReceive(message); // Is it strange that I call this implicitly?
-        receive(port, message);
-    }
-
-    private void recordDone(ActorNode.Port port) {
-        if (activePorts.contains(port)) activePorts.remove(port);
-        else if (pendingPorts.contains(port)) pendingPorts.remove(port);
-        else throw TypeDBException.of(ILLEGAL_STATE);
-    }
-
-    @Override
-    protected void exception(Throwable e) {
-        nodeRegistry.terminate(e);
-    }
-
-    @Override
-    public void terminate(Throwable e) {
-        super.terminate(e);
     }
 
     public static class Port {
@@ -157,7 +74,7 @@ public abstract class ActorNode<NODE extends ActorNode<NODE>> extends Actor<NODE
         private int lastRequestedIndex;
         private boolean isPending;
 
-        private Port(ActorNode<?> owner, ActorNode<?> remote) {
+        protected Port(ActorNode<?> owner, ActorNode<?> remote) {
             this.owner = owner;
             this.remote = remote;
             this.state = State.READY;
@@ -165,7 +82,7 @@ public abstract class ActorNode<NODE extends ActorNode<NODE>> extends Actor<NODE
             this.isPending = false;
         }
 
-        private void recordReceive(Message msg) {
+        protected void recordReceive(Message msg) {
             assert state == State.PULLING;
             if (msg.type() == Message.MessageType.SNAPSHOT) { // TODO
                 throw TypeDBException.of(UNIMPLEMENTED);
@@ -174,7 +91,6 @@ public abstract class ActorNode<NODE extends ActorNode<NODE>> extends Actor<NODE
             assert lastRequestedIndex == msg.index();
             if (msg.type() == Message.MessageType.DONE) {
                 state = State.DONE;
-                owner.recordDone(this);
             } else if (msg.type() == Message.MessageType.SNAPSHOT) {
                 this.isPending = true;
                 state = State.READY;
