+++ reasoner/v4/nodes/ConclusionNode.java
@@ -5,9 +5,7 @@ import com.vaticle.typedb.core.common.iterator.FunctionalIterator;
 import com.vaticle.typedb.core.concept.answer.ConceptMap;
 import com.vaticle.typedb.core.logic.Rule;
 import com.vaticle.typedb.core.reasoner.controller.ConjunctionController;
-import com.vaticle.typedb.core.reasoner.v4.ActorNode;
 import com.vaticle.typedb.core.reasoner.v4.Message;
-import com.vaticle.typedb.core.reasoner.v4.NodeRegistry;
 
 import java.util.Optional;
